requirements.txt
asyncio
beautifulsoup4
discord.py
gspread
oauth2client==4.1.2
psycopg2-binary
requests